Namespace renamed TokenType > TokenTypes
