--- src/TokenType/AbstractTokenType.php
@@ -1,75 +1,0 @@
-<?php
-/**
- * OAuth 2.0 Abstract Token Type
- *
- * @package     league/oauth2-server
- * @author      Alex Bilbie <hello@alexbilbie.com>
- * @copyright   Copyright (c) Alex Bilbie
- * @license     http://mit-license.org/
- * @link        https://github.com/thephpleague/oauth2-server
- */
-
-namespace League\OAuth2\Server\TokenType;
-
-use League\OAuth2\Server\AbstractServer;
-use League\OAuth2\Server\Entity\SessionEntity;
-
-abstract class AbstractTokenType
-{
-    /**
-     * Response array
-     *
-     * @var array
-     */
-    protected $response = [];
-
-    /**
-     * Server
-     *
-     * @var \League\OAuth2\Server\AbstractServer $server
-     */
-    protected $server;
-
-    /**
-     * Server
-     *
-     * @var \League\OAuth2\Server\Entity\SessionEntity $session
-     */
-    protected $session;
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setServer(AbstractServer $server)
-    {
-        $this->server = $server;
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setSession(SessionEntity $session)
-    {
-        $this->session = $session;
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setParam($key, $value)
-    {
-        $this->response[$key] = $value;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getParam($key)
-    {
-        return isset($this->response[$key]) ? $this->response[$key] : null;
-    }
-}
